Further implementation
attack is a locked state. bc of the way movement is done atm you cannot move while attacking (in theory) in practise this is unnoticeable bc the attack state exits as soon as it enters.
If however, in the current system, we made an attack anim and had to wait for the attack to be 'done' the character would be standing still during this time, except for the momentum of the rigidbody. /would not be receiving movement inputs while in the attack state.